+++ envconfig/src/traits.rs
@@ -3,7 +3,17 @@ use crate::error::Error;
 /// Indicates that structure can be initialize from environment variables.
 pub trait Envconfig {
     /// Initialize structure from environment variables.
+    /// Deprecated in favor of init_from_env().
+    #[deprecated(
+        since = "0.9.0",
+        note = "Function init() is deprecated. Please use init_from_env() instead."
+    )]
     fn init() -> Result<Self, Error>
+    where
+        Self: Sized;
+
+    /// Initialize structure from environment variables.
+    fn init_from_env() -> Result<Self, Error>
     where
         Self: Sized;
 }