+++ test_suite/tests/neasting.rs
@@ -35,7 +35,7 @@ fn test_neasting() {
     env::set_var("DB_HOST", "localhost");
     env::set_var("DB_PORT", "5432");
 
-    let config = Config::init().unwrap();
+    let config = Config::init_from_env().unwrap();
     assert_eq!(config.db.host, "localhost");
     assert_eq!(config.db.port, 5432u16);
 }
@@ -46,7 +46,7 @@ fn test_neasting_error() {
 
     env::set_var("DB_HOST", "localhost");
 
-    let err = Config::init().err().unwrap();
+    let err = Config::init_from_env().err().unwrap();
     let expected_err = Error::EnvVarMissing { name: "DB_PORT" };
     assert_eq!(err, expected_err);
 }
@@ -58,7 +58,7 @@ fn test_duplicated_are_allowed() {
     env::set_var("DB_HOST", "localhost");
     env::set_var("DB_PORT", "5432");
 
-    let config = ConfigDouble::init().unwrap();
+    let config = ConfigDouble::init_from_env().unwrap();
     assert_eq!(config.db1.host, "localhost");
     assert_eq!(config.db1.port, 5432u16);
     assert_eq!(config.db2.host, "localhost");